change navbar styling and add new vision on home page,move some files in correct folder
