+++ src/components/Modal.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import styles from '../styles/Modal.module.css';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faXmark } from '@fortawesome/free-solid-svg-icons';
 
 interface ModalProps {
     isOpen: boolean;
@@ -13,7 +15,7 @@ const Modal: React.FC<ModalProps> = ({ isOpen, onClose, children }) => {
     return (
         <div className={styles.overlay}>
             <div className={styles.modal}>
-            <button style={{ backgroundColor: '#ffffff'}} className={styles.closeBtn} onClick={onClose}>X</button>
+            <button style={{ backgroundColor: '#ffffff'}} className={styles.closeBtn} onClick={onClose}><FontAwesomeIcon icon={faXmark}/></button>
 
 
                 {children}
