+++ src/middleware.ts
@@ -7,5 +7,5 @@ export default withAuth({
 })
 
 export const config = {
-    matcher: ['/', '/createCategory']
+    matcher: ['/', '/categories']
 }
