+++ src/styles/Category.module.css
@@ -15,7 +15,7 @@
 .category {
     display: flex;
     flex-direction: column;
-    background-color: #eee;
+    background-color: #f2f6fc;;
     padding: 30px 0;
     width: 100%;
     align-items: center;
@@ -49,7 +49,7 @@
     display: flex;
     flex-direction: row;
     flex-wrap: wrap;
-    background-color: #eee;
+    background-color:#f2f6fc;
     padding: 30px 0;
     width: 100%;
     justify-content: center;
@@ -62,6 +62,7 @@
 .categoryName {
     width: 28%;
     padding: 20px;
+    background-color:#fffdfd;
     box-shadow: 0 0 5px rgba(0, 0, 0, 0.1);
     border-radius: 8px;
     margin-bottom: 2rem;
